Implement interfaces for publicroomsapi storage
